Fix "Warning: Cannot update a component from inside the function body of a different component."
I can't really remember why I added this sort of "un-initialisation" step, but it seems irrelevant to set to none a thing set right after - and more importantly, it's creating rendering issues!
Since calendars event are kind of cached, this doesn't make sens now that we have a cached context and now a hook.